+++ woocommerce-gateway-ebanx/services/class-wc-ebanx-constants.php
@@ -139,6 +139,20 @@ abstract class WC_EBANX_Constants
 	);
 
 	/**
+	 * The cash payments processed by EBANX
+	 *
+	 * @var array
+	 */
+	public static $CASH_PAYMENTS_GATEWAYS_CODE = array(
+		'ebanx-banking-ticket',
+		'ebanx-oxxo',
+		'ebanx-pagoefectivo',
+		'ebanx-sencillito',
+		'ebanx-safetypay-cash',
+		'ebanx-baloto',
+	);
+
+	/**
 	 * The banks that EBANX process in Brazil
 	 *
 	 * @var array